feat: implement farmerForm model for mongoDB
